Attach multiple organizations to a contact
This could be useful if a contacts represents more than 1 organization (i.e. one single contact for 2 clients).

The original "Business info" section in the contact CRUD form and the "organization" was removed. This section was ow replaced by an "Organizations" table in the CRUD form page.

The "Organization name and job title" column in the contacts CRUD table was also removed.

References:
- https://filamentphp.com/docs/3.x/panels/resources/relation-managers